Refactored Health Check schemas into the view logic
I think I'm going to keep schema definitions that are only used by a single endpoint right next to the view code. I'm jumping around through the codebase way more than seems strictly necessary.
